Import live.2005.04.01.tar.gz
2005.03.23
- Changed "BasicTaskScheduler::SingleStep()" to call only one read handler during
  each iteration of the event loop.  This prevents potential problems if
  the event loop is called reentrantly from within a read handler.
- Added a new, alternative version of "ByteStreamFileSource::createNew()" that
  takes an open file (FILE*) as parameter, instead of a file name.
  (Thanks to Mike Yan for this suggestion.)
- Updated the top-level Makefile to build the lower-level directories using
  "$(MAKE)" rather than "make".  This allows a command other than "make"
  (e.g., "gmake") to be used.  (Thanks to Sergio.Gelato for this fix.)
- Minor changes to (apparently) make Sun's C++ compiler happy.
  (Thanks to Sergio.Gelato for these.)
- Some changes to "RTSPServer" to make it more tolerant of some strange, non-standard
  clients.
- Support seeking within RTSP/RTP-streamed MPEG-1 or 2 Program Stream files.

2005.03.28
- The "RTSPServer" implementation now supports raw-UDP streaming, for those
  clients that request it.  (This is usually requested only by some specialized
  clients - such as set-top boxes - and only for data such as MPEG Transport
  Streams where audio/video is muxed together (and no RTP timestamps are needed).)
- Updated the calculation of MP3 sampling frequencies to allow for 'MPEG2.5'
  (Thanks to Massimo Buffo for this fix.)

2005.03.30
- Improved the support for seeking within RTSP/RTP-streamed MPEG-1 or 2
  Program Stream files.  (A/V sync after a seek still isn't perfect, though.)

2005.03.31
- Removed some testing code that had accidentally been left in the previous
  release.  If you use "openRTSP", then please use this release instead of the
  previous one.
- Fixed a bug in "RTPServer" when handling streams without stream names.
  (The bug had been introduced in version 2005.03.28.)

2005.04.01
No comments

+++ liveMedia/ServerMediaSession.cpp
@@ -111,7 +111,7 @@ void ServerMediaSession::testScaleFactor(float& scale) {
     subsession->testScaleFactor(ssscale);
     if (subsession == fSubsessionsHead) { // this is the first subsession
       minSSScale = maxSSScale = bestSSScale = ssscale;
-      bestDistanceTo1 = fabsf(ssscale - 1.0f);
+      bestDistanceTo1 = (float)fabs(ssscale - 1.0f);
     } else {
       if (ssscale < minSSScale) {
 	minSSScale = ssscale;
@@ -119,7 +119,7 @@ void ServerMediaSession::testScaleFactor(float& scale) {
 	maxSSScale = ssscale;
       }
 
-      float distanceTo1 = fabsf(ssscale - 1.0f);
+      float distanceTo1 = (float)fabs(ssscale - 1.0f);
       if (distanceTo1 < bestDistanceTo1) {
 	bestSSScale = ssscale;
 	bestDistanceTo1 = distanceTo1;
